moved search modal to itsown component
